Move 404/500 test views to homepage.views.diagnostics
