+++ homepage/views/diagnostics.py
